summary
This project will allow users to create an account. Save that account to a local storage and then later 
when entered properly take them to a page with a picture of a kitten on it. If I feel incredibly ambitious I will 
have the program send a confirmation email to the user.